% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pgb.R
\name{theme_pgb}
\alias{theme_pgb}
\title{Customized ggplot2 theme}
\usage{
theme_pgb()
}
\value{
A ggplot2 theme object.
}
\description{
This function provides a customized ggplot2 theme with specific text, axis, grid, and legend styles.
It is designed to be consistent with the visual preferences for PGB plots.
}
